requirements.txt
requests
atproto
emoji
IPython
python-dotenv
docopt